working on entity interpolation

+++ src/server/Enemy.java
@@ -3,7 +3,7 @@ package server;
 import server.npc.NPC;
 
 public class Enemy extends NPC {
-    public Enemy(String name, float spawnX, float spawnY) {
+    public Enemy(String name, double spawnX, double spawnY) {
         super(name, spawnX, spawnY);
     }
 }
